+++ ukrainian_integrations/pbx_sms/vitalpbx/custom_fields.py
@@ -86,7 +86,9 @@ def ensure_integration_custom_fields():
             {"fieldname": "liqpay_paid_amount", "label": "LiqPay Paid Amount", "fieldtype": "Currency", "insert_after": "liqpay_transaction_id", "read_only": 1, "no_copy": 1},
         ],
         "Sales Order": [
-            {"fieldname": "ua_external_order_key", "label": "UA External Order Key", "fieldtype": "Data", "insert_after": "po_no", "read_only": 1, "unique": 1, "no_copy": 1, "length": 140}
+            {"fieldname": "ua_external_order_key", "label": "UA External Order Key", "fieldtype": "Data", "insert_after": "po_no", "read_only": 1, "unique": 1, "no_copy": 1, "length": 140},
+            {"fieldname": "ua_ecommerce_channel", "label": "Ecommerce Channel", "fieldtype": "Link", "options": "Ecommerce Channel", "insert_after": "ua_external_order_key", "read_only": 1, "no_copy": 1},
+            {"fieldname": "ua_external_order_id", "label": "External Order ID", "fieldtype": "Data", "insert_after": "ua_ecommerce_channel", "read_only": 1, "no_copy": 1, "length": 140}
         ],
         "Customer": [
             {"fieldname": "ua_external_customer_key", "label": "UA External Customer Key", "fieldtype": "Data", "insert_after": "customer_name", "read_only": 1, "unique": 1, "no_copy": 1, "length": 80}